Okay we are slowly trying to implement labelled number types
i8, i16, etc.  Just need to massage them into the lexer and parser.

+++ src/hir.rs
@@ -117,7 +117,7 @@ impl<T> Expr<T> {
     }
 
     /// Shortcut function for making literal integers
-    pub const fn int(i: i64) -> Self {
+    pub const fn int(i: i128) -> Self {
         Self::Lit {
             val: Literal::Integer(i),
         }
